requirements.txt
django==4.2.5
djangorestframework==3.14.0
django-rest-framework==0.1.0
adrf==0.1.2 
async-property==0.2.2
psycopg==3.1.12
psycopg2==2.9.8
psycopg-pool==3.1.8
psycopg-binary==3.1.12